git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/bamRecursion@12537 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/cpa/octagon/coefficients/OctNumericValue.java
@@ -47,7 +47,7 @@ public class OctNumericValue {
     this.intVal = BigInteger.valueOf(intVal);
   }
 
-  public OctNumericValue(float floatVal) {
+  public OctNumericValue(double floatVal) {
     this.floatVal = BigDecimal.valueOf(floatVal);
   }
